+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # 更新日志
 
+## 4.0.2
+- 适配honoka-utils 1.0.6。
+- 适配qqrobot-spring-boot-starter 1.2.1。
+
 ## 4.0.1
 - 适配新版honoka-utils和qqrobot-spring-boot-starter。